+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+### Added
+- `nextAdBreak()` to `AdTimeline`
+- `previousAdBreak()` to `AdTimeline`
+
 ### Changed
 - Bitmovin player to `2.58.0+jason`